python_web/payment/pyproject.toml
[project]
name = "payment"
version = "0.1.0"
description = "Payment & Orders microservice"
requires-python = "==3.12.*"
dependencies = [
    "backend-common",
    "fastapi>=0.129.0",
    "asyncpg>=0.31.0",
    "pydantic-settings>=2.13.0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.46",
    "aiokafka>=0.10.0",
    "uvicorn>=0.41.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "cryptography>=44.0.0",
    "alembic>=1.15.0",
    "ruff>=0.15.10",
]

[tool.uv.sources]
backend-common = { path = "../../shared", editable = true }

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.27",
    "aiosqlite>=0.22",
    "pytest-cov>=7.1.0",
]
